test: fiks flaky test ved å flytte mockOAuth2Server til AbstractMockkSpringRunner
TokenContextHolder er et JVM-singleton som ble overskrevet av
AbstractSpringIntegrationTest sin Spring-kontekst (TOKEN_X_ISSUER=dummy),
noe som fikk hasTokenFor("tokenx") til å returnere false og controleren
til å svare 403 i stedet for 400 når token inneholder ugyldig fødselsnummer.

Fikset ved å flytte mockOAuth2Server og @DynamicPropertySource til
AbstractMockkSpringRunner slik at alle Spring-kontekster i JVM-en alltid
får riktig issuer-mapping uavhengig av opprettelsesrekkefølge.
